requirements.txt
azure_storage==0.37.0
librosa==0.10.0
numpy==1.21.5
openai==0.26.5
pyttsx3==2.90
scipy==1.8.0
sounddevice==0.4.5
soundfile==0.12.1
speechbrain==0.5.13
streamlit==1.16.0
torch==1.13.1
transformers==4.26.1
